refactor: restructure codebase to follow Feature-Sliced Design (FSD)
- Add entities/game layer with model (types, games data) and ui (GameCard, GamesGrid) components
- Add features/game-selection layer for game selection business logic
- Remove duplicate games data and fix incorrect links
- Clean up shared layer by moving Game types to entities
- Update HomePage to use new FSD structure
- Remove old widget components that moved to entities
- Update AGENTS.md with FSD architecture guidelines
