+++ src/entities/game/index.ts
@@ -1,0 +1,4 @@
+export { GameCard } from "./ui/GameCard";
+export { GamesGrid } from "./ui/GamesGrid";
+export type { Game } from "./model/types";
+export { games } from "./model/games";
