+++ src/entities/game/model/types.d.ts
