+++ src/features/game-selection/index.ts
@@ -1,0 +1,1 @@
+export { GameSelection } from "./ui/GameSelection";
