+++ src/features/game-selection/ui/GameSelection.tsx
@@ -1,0 +1,5 @@
+import { GamesGrid, games } from "~/entities/game";
+
+export function GameSelection() {
+  return <GamesGrid games={games} />;
+}
