+++ src/shared/types/index.d.ts
@@ -1,4 +1,3 @@
 import type { DifficultyNamedSizes, Size, NumberNamedSizes } from "./size";
-import type { Game, Difficulty } from "./game";
 
-export { Difficulty, Game, Size, NumberNamedSizes, DifficultyNamedSizes };
+export { Size, NumberNamedSizes, DifficultyNamedSizes };
