--- src/widgets/app-footer.tsx
@@ -1,9 +1,0 @@
-function AppFooter() {
-  return (
-    <footer className="flex items-center justify-center border-t border-zinc-300 bg-zinc-100 p-4">
-      <p>Levchenko Artem</p>
-    </footer>
-  );
-}
-
-export { AppFooter };
